basename works
