+++ web-frontend/src/client.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { components } from "./openapi_schemas";
 
-const apiURL = import.meta.env.REACT_APP_API_PREFIX || "../api/v1";
+const apiURL = import.meta.env.REACT_APP_API_PREFIX || "../../api/v1";
 
 var axiosInstance = axios.create({
   baseURL: apiURL,
