+++ web-frontend/vite.config.js
@@ -3,6 +3,7 @@ import react from "@vitejs/plugin-react"
 import { viteRequire } from "vite-require"
 
 export default defineConfig({
+    base: "/ui/",
     server: {
         proxy: {
             "/api": {
